% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo_test.R
\name{hypo_test}
\alias{hypo_test}
\title{Perform Hypothesis Testing on Model Accuracy}
\usage{
hypo_test(
  model,
  data,
  label,
  null.value = 0.8,
  alternative = c("two sided", "less", "greater"),
  alpha = 0.05,
  method = c("z", "lrt")
)
}
\arguments{
\item{model}{An SVM model list object returned by \code{svm()}.}

\item{data}{A data frame of numerical features data.}

\item{label}{A label vector of 1 or -1 with respect to data.}

\item{null.value}{Null hypothesis of accuracy value. Defaults to 0.8}

\item{alternative}{Type of alternative hypothesis. Two sided, greater, or less.}

\item{alpha}{Significance level. Defaults to 0.05}

\item{method}{Type of test to use. Z-test or LRT.}
}
\value{
A list containing:
\describe{
  \item{estimate}{Estimated accuracy of the model.}
  \item{n}{Sample size.}
  \item{null.value}{Null hypothesis value.}
  \item{alternative}{Type of alternative hypothesis.}
  \item{method}{Test method. Z-test or LRT.}
  \item{alpha}{Significance level.}
  \item{test.statistic}{Test statistic value.}
  \item{p.value}{P-value.}
  \item{status}{Rejection status of the null hypothesis. Reject or fail to reject.}
  \item{conf.int}{Confidence interval (only for Z-test).}
}
}
\description{
Conducts a one-sample proportion Z-test or likelihood ratio test (LRT) to assess if the model's accuracy differs significantly from a
specified null value.
}
\examples{
data = data.frame(height = c(169, 179, 159, 165, 190), weight = c(70, 80, 85, 75, 70))
label = c(1, -1, 1, -1, -1)
model = svm(data, label)
hypo_test(model, data, label, null.value = 0.5, alternative = "less")

}
