% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\title{Kernel Function}
\usage{
kernel(Xi, Xj, type, degree = 2, c = 1, gamma = 1)
}
\arguments{
\item{Xi}{First vector.}

\item{Xj}{Second vector.}

\item{type}{Kernel type polynomial or RBF.}

\item{degree}{Degree of the polynomial kernel.}

\item{c}{A constant for polynomial kernel.}

\item{gamma}{Parameter for RBF kernel.}
}
\value{
A kernel value.
}
\description{
Calculate value for polynomial and RBF kernel.
}
\keyword{internal}
