% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot Histogram of Data}
\usage{
plot_hist(data, column = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{column}{The column index or name of data to plot.}
}
\value{
A neat histogram plot of a column from data.
}
\description{
Plot a histogram to provide an overview of the data distribution.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
plot_hist(data, "height")

}
