% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict Label}
\usage{
predict(model, data)
}
\arguments{
\item{model}{An SVM model list object returned by \code{svm()}.}

\item{data}{A data frame of numerical features data.}
}
\value{
A vector of -1 and 1 that represents the label prediction.
}
\description{
Classify data based on the the trained SVM model's kernel, features, and variables.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
label = c(-1, -1, 1, 1, -1)
linear.model = svm(data, label)
test.data = data.frame(height = c(175, 195, 155, 165, 185), weight = c(75, 85, 80, 70, 75))
test.label = data.frame(-1, -1, 1, -1, -1)
predict(linear.model, test.data)

}
