% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Data and Label for Binary SVM Classifier}
\usage{
preprocess(data, transform = NULL, label.col = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{transform}{A list mapping column names to numeric levels for converting character features to numeric.}

\item{label.col}{An index or name of column of the binary label.}
}
\value{
A data frame with numerical feature columns and a -1 or 1 label column.
}
\description{
Processes a data frame by:
  1. Converting character features into numeric levels based on the provided mappings,
  2. Removing rows containing NA values,
  3. Dropping character columns without a mapping, and
  4. Setting the binary label to -1 and 1.
}
\examples{
data = data.frame(weather = c("Rainy", "Sunny", "Sunny"),
                  temp = c("Cold", "Mild", "Hot"), play = c(1, 0, 1))
transform = list(weather = c("Rainy", "Sunny"), temp = c("Cold", "Mild", "Hot"))
data = preprocess(data, transform, 3)

}
