% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_correlated_features.R
\name{select_correlated_features}
\alias{select_correlated_features}
\title{Select Correlated Features}
\usage{
select_correlated_features(data, label, n = NULL, min = 0.3, show = FALSE)
}
\arguments{
\item{data}{A data frame of numerical features data.}

\item{label}{A label vector of 1 or -1 with respect to data.}

\item{n}{Number of best features to select.}

\item{min}{Minimum correlation for a feature to be selected. Defaults to 0.3}

\item{show}{Print the sorted absolute correlations.}
}
\value{
A vector of n most meaningful features, or a vector of features with absolute correlation above 0.2 with the label, if n is not specified.
}
\description{
Choose features with higher correlations with label.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
label = c(-1, -1, 1, 1, -1)
select_correlated_features(data, label, 1)

}
