% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{svm}
\alias{svm}
\title{SVM Model Training}
\usage{
svm(
  data,
  label,
  n = nrow(data),
  features = names(data),
  type = "linear",
  C = 1,
  degree = 2,
  c = 1,
  gamma = 1/length(features)
)
}
\arguments{
\item{data}{A data frame of numerical features data.}

\item{label}{A label vector of 1 or -1 with respect to data.}

\item{n}{Number of data to use for training. Defaults to the number of rows in data.}

\item{features}{A vector of column names of features to use for training. Defaults to all column names.}

\item{type}{Kernel type to use for training. Only linear, polynomial, and RBF kernels are available. Defaults to linear.}

\item{C}{Regularization parameter for quadratic programming. Defaults to 1.}

\item{degree}{Degree of the polynomial kernel. Defaults to 2.}

\item{c}{Constant term for polynomial kernel. Defaults to 1.}

\item{gamma}{Parameter for RBF kernel. Defaults to 1 / number of features.}
}
\value{
A list that represents the SVM model:
\describe{
  \item{w}{Weight vector for linear kernel.}
  \item{alpha}{Dual solution of the quadratic programming.}
  \item{X}{Scaled data used in training}
  \item{y}{Label used in training}
  \item{b}{Bias term.}
  \item{type}{Kernel type used in training}
  \item{features}{Features used in training.}
  \item{n}{Number of data used in training.}
  \item{c}{Constant used in the polynomial kernel.}
  \item{C}{Regularization parameter used in quadratic programming (dual).}
  \item{degree}{Degree used in the polynomial kernel.}
  \item{gamma}{Gamma used in the RBF kernel.}
}
}
\description{
Train SVM binary classifier for numeric features data using quadratic programming. Supports linear, polynomial, and RBF kernel types.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
label = c(-1, -1, 1, 1, -1)
linear.model = svm(data, label)

}
