% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm01xht.R
\name{svm01xht}
\alias{svm01xht}
\title{svm01xht: Support Vector Machine (SVM) for Binary Classification Model with Hypothesis Testing Evaluation}
\description{
The svm01xht package provides functionality for:
\itemize{
 \item Structured binary support vector machines (SVM) implementation
 \item Hypothesis testing of models' accuracy.
}
}
\section{Binary SVM implementation}{

\itemize{
 \item The `plot_hist()` function allows users to visualize distribution of data
 \item The `preprocess()` function allows users to convert character data to numeric and clean the data for SVM implementation.
 \item The `select_correlated_features()` function allows users to select meaningful features by analyzing the correlations between features and the binary label.
 \item The `svm()` function allows users to train a binary SVM model with linear, polynomial, or RBF kernels.
 \item The `predict()` function allows users to use the trained SVM model to predict label from feature data.
 \item The `test_accuracy()` function allows users to evaluate the accuracy between prediction and actual label.
 \item The `tune_parameters()` function allows users to compare performance of different parameters for the SVM model.
}
}

\section{Accuracy Evaluation through Hypothesis Testing}{

\itemize{
 \item The `hypo_test()` function allows users to conduct Z-test or likelihood ratio test (LRT) on the models' accuracy.
 \item The `visualize_ht()` visualizes the hypothesis testing result from `hypo_test()` to help users understand the test result.
}
}

