% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_accuracy.R
\name{test_accuracy}
\alias{test_accuracy}
\title{Test Model's Accuracy}
\usage{
test_accuracy(model, data, label)
}
\arguments{
\item{model}{An SVM model list object returned by \code{svm()}.}

\item{data}{A data frame of numerical features data.}

\item{label}{A label vector of 1 or -1 with respect to data.}
}
\value{
The level of accuracy from 0 to 1.
}
\description{
Compare predicted values with actual values to calculate model's accuracy.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
label = c(-1, -1, 1, 1, -1)
linear.model = svm(data, label)
test.data = data.frame(height = c(175, 195, 155, 165, 185), weight = c(75, 85, 80, 70, 75))
test.label = data.frame(-1, -1, 1, -1, -1)
test_accuracy(linear.model, test.data, test.label)

}
