% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_parameters.R
\name{tune_parameters}
\alias{tune_parameters}
\title{Tune SVM Parameters}
\usage{
tune_parameters(
  training.data,
  training.label,
  test.data,
  test.label,
  features = names(training.data),
  type = "linear",
  C = 1,
  degree = 2,
  c = 1,
  gamma = 1/length(features)
)
}
\arguments{
\item{training.data}{A data frame of numerical features data to be used for training.}

\item{training.label}{A label vector of 1 or -1 with respect to training.data.}

\item{test.data}{A data frame of numerical features data to be used for testing.}

\item{test.label}{A label vector of 1 or -1 with respect to test.data.}

\item{features}{A vector of column names of features to use for training. Defaults to all column names.}

\item{type}{Kernel type to use for training. Only linear, polynomial, and RBF kernels are available. Defaults to linear.}

\item{C}{Regularization parameters (vector) for quadratic programming to test. Defaults to 1.}

\item{degree}{Degrees (vector) of the polynomial kernel to test. Defaults to 2.}

\item{c}{Constants (vector) for polynomial kernel to test. Defaults to 1.}

\item{gamma}{Parameters (vector) for RBF kernel to test. Defaults to 1 / number of features.}
}
\value{
A data frame showing the of training accuracy and test accuracy for each combination of parameters.
}
\description{
Create several SVM models with same training data and test data, but different parameters to determine better parameters.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
test.data = data.frame(height = c(175, 195, 155, 165, 185), weight = c(75, 85, 80, 70, 75))
label = c(-1, -1, 1, 1, -1)
test.label = c(-1, -1, 1, 1, -1)
tune_parameters(data, label, test.data, test.label, type = "linear", C = c(0.1, 1, 10))

}
