% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ht.R
\name{visualize_ht}
\alias{visualize_ht}
\title{Visualize Hypothesis Testing Results}
\usage{
visualize_ht(test.result, type = c("I", "II"))
}
\arguments{
\item{test.result}{A list of hypothesis testing result from hypo_test().}

\item{type}{Type I or II error to plot. Type II error plot is only available for Z-test.}
}
\value{
A plot showing either:
\itemize{
  \item For type = "I": The statistic and rejection region of Z-test or LRT.
  \item For type = "II": Power with respect to possible true alternatives.
}
}
\description{
Creates plots for visualizing type I error of Z-test and LRT, or type II error, but only for Z-test.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
label = c(-1, -1, 1, 1, -1)
model = svm(data, label)
test = hypo_test(model, data, label)
visualize_ht(test)

}
