% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\title{Kernel Function}
\usage{
kernel(Xi, Xj, type, degree = 2, c = 1, gamma = 1)
}
\arguments{
\item{Xi}{First input vector.}

\item{Xj}{Second input vector.}

\item{type}{Kernel type. Polynomial or RBF.}

\item{degree}{Degree of the polynomial kernel.}

\item{c}{Constant term for the polynomial kernel.}

\item{gamma}{Parameter for the RBF kernel.}
}
\value{
The kernel value.
}
\description{
Calculate value for polynomial and RBF kernel.
}
\keyword{internal}
