% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict Label}
\usage{
predict(model, data)
}
\arguments{
\item{model}{A list that represents an SVM model from svm().}

\item{data}{A data frame of numerical features data.}
}
\value{
A vector of -1 and 1 that represents the label prediction.
}
\description{
Classify data based on the the trained SVM model's kernel, features, and variables.
}
\examples{
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
test.data = data.frame(height = c(175, 195, 155, 165, 185), weight = c(75, 85, 80, 70, 75))
model = list(w = c(-1.58, 1.3), b = 0.2, type = "linear", features = c("height", "weight"))
names(model$w) = c("height", "weight")
predict(model, test.data)

}
