% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Data and Label for Binary SVM Classifier}
\usage{
preprocess(data, transform = NULL, label.col = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{transform}{A list which maps column names to levels for character into numeric feature conversion.}

\item{label.col}{An index or name of column of the binary label.}
}
\value{
A clean numerical data frame of features and label.
}
\description{
Process data by mapping character features to numeric features, removing rows with NA values,
removing character feature columns that were not mapped, and adjusting label from binary to 1 and -1.
}
\examples{
data = data.frame(weather = c("Rainy", "Sunny", "Sunny"),
                  temp = c("Cold", "Mild", "Hot"), play = c(1, 0, 1))
transform = list(weather = c("Rainy", "Sunny"), temp = c("Cold", "Mild", "Hot"))
data = preprocess(data, transform, 3)

}
