% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_accuracy.R
\name{test_accuracy}
\alias{test_accuracy}
\title{Test Model's Accuracy}
\usage{
test_accuracy(model, data, label)
}
\arguments{
\item{model}{A list that represents an SVM model from svm().}

\item{data}{A data frame of numerical features data.}

\item{label}{A label vector of 1 or -1 with respect to data.}
}
\value{
A number between 0 and 1 that represents level of accuracy.
}
\description{
Compare predicted values with actual values to evaluate accuracy of a model.
}
\examples{
model = list(w = c(-1.58, 1.3), b = 0.2, type = "linear", features = c("height", "weight"))
data = data.frame(height = c(170, 190, 150, 160, 180), weight = c(70, 80, 85, 75, 70))
names(model$w) = c("height", "weight")
label =  c(-1, -1, 1, 1, -1)
test_accuracy(model, data, label)

}
